codacy: remove duplicate code.
